library system/requirements.txt
mysql-connector-python==8.2.0
qrcode[pil]==7.4.2
Pillow==10.1.0
